[KYV-1329] Use HDS header to get HDS language selector
