+++ src/i18n/en.json
@@ -8,12 +8,12 @@
       "fi": {
         "code-uppercase": "FI",
         "code": "fi",
-        "label": "Finnish"
+        "label": "Suomi"
       },
       "sv": {
         "code-uppercase": "SV",
         "code": "sv",
-        "label": "Swedish"
+        "label": "Svenska"
       },
       "en": {
         "code-uppercase": "EN",
